+++ src/Elders.Pandora/IConfigurationRepository.cs
@@ -1,0 +1,8 @@
+﻿namespace Elders.Pandora
+{
+    public interface IConfigurationRepository
+    {
+        string Get(string key);
+        void Set(string key, string value);
+    }
+}